refactor(render): move the frame's leaves into their own packages
Eleven files of render/ read nothing of the frame and were read by it
alone or by the mixins: the material maps and their atlases, the storage
images and buffers with the encoder work that creating or freeing either
records, the per-pass timings with the ring timings and the wide-wait
switch, and the module cache, a thousand lines with no line in common
with the frame. They go to render.pbr, render.storage, render.timing and
cache, a package note on each saying what it is a leaf of. What the
frame still reads of them opens to it, and a note that linked a class
of the frame from a leaf names it instead.
