+++ common/src/main/java/dev/vitrail/mixin/CommandEncoderMixin.java
@@ -1,7 +1,7 @@
 package dev.vitrail.mixin;
 
 import dev.vitrail.render.GeometryHold;
-import dev.vitrail.render.PassTimings;
+import dev.vitrail.render.timing.PassTimings;
 
 import com.mojang.blaze3d.systems.CommandEncoder;
 import com.mojang.blaze3d.systems.RenderPass;
