+++ common/src/main/java/dev/vitrail/mixin/GlslCompilerMixin.java
@@ -6,9 +6,10 @@ import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.mojang.blaze3d.shaders.ShaderType;
 import com.mojang.blaze3d.vulkan.glsl.GlslCompiler;
 import com.mojang.blaze3d.vulkan.glsl.IntermediaryShaderModule;
+import dev.vitrail.cache.ModuleCache;
 import dev.vitrail.glsl.LoadClock;
-import dev.vitrail.render.ModuleCache;
 import dev.vitrail.render.RawLocals;
+import dev.vitrail.render.storage.StorageImages;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Coerce;
