+++ common/src/main/java/dev/vitrail/mixin/VulkanBindGroupLayoutMixin.java
@@ -5,8 +5,8 @@ import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.mojang.blaze3d.vulkan.VulkanBindGroupLayout;
 import com.mojang.blaze3d.vulkan.VulkanBindGroupLayout.Entry;
 import dev.vitrail.pack.texture.CustomImages;
-import dev.vitrail.render.StorageBuffers;
-import dev.vitrail.render.StorageImages;
+import dev.vitrail.render.storage.StorageBuffers;
+import dev.vitrail.render.storage.StorageImages;
 import org.lwjgl.vulkan.VkDescriptorSetLayoutBinding;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
