+++ common/src/main/java/dev/vitrail/mixin/VulkanQueueSubmitMixin.java
@@ -1,6 +1,6 @@
 package dev.vitrail.mixin;
 
-import dev.vitrail.render.PassTimings;
+import dev.vitrail.render.timing.PassTimings;
 
 import com.mojang.blaze3d.vulkan.VulkanQueue;
 import org.spongepowered.asm.mixin.Mixin;
