+++ common/src/main/java/dev/vitrail/render/GeometryProgram.java
@@ -7,12 +7,17 @@ import dev.vitrail.mixin.access.GpuDeviceAccessor;
 import dev.vitrail.pack.model.AlphaTest;
 import dev.vitrail.pack.model.ProgramStage;
 import dev.vitrail.pack.model.RenderStage;
+import dev.vitrail.pack.model.TargetName;
+import dev.vitrail.pack.model.TextureStage;
 import dev.vitrail.pack.program.TerrainPass;
 import dev.vitrail.pack.target.ChainPlan;
 import dev.vitrail.pack.target.DrawBuffers;
 import dev.vitrail.pack.target.SamplerPlan;
-import dev.vitrail.pack.model.TargetName;
-import dev.vitrail.pack.model.TextureStage;
+import dev.vitrail.render.pbr.PbrAtlases;
+import dev.vitrail.render.pbr.PbrMap;
+import dev.vitrail.render.pbr.PbrTextures;
+import dev.vitrail.render.storage.StorageBuffers;
+import dev.vitrail.render.timing.PassTimings;
 import dev.vitrail.uniform.ClipSpace;
 import dev.vitrail.uniform.TextSink;
 import dev.vitrail.uniform.WorldState;
