+++ common/src/main/java/dev/vitrail/render/PackChoice.java
@@ -14,6 +14,10 @@ import dev.vitrail.pack.source.ShaderPackSource;
 import dev.vitrail.pack.target.PackDirectives;
 import dev.vitrail.pack.texture.CustomImages;
 import dev.vitrail.pack.texture.CustomStorage;
+import dev.vitrail.render.pbr.PbrAtlases;
+import dev.vitrail.render.storage.StorageBuffers;
+import dev.vitrail.render.storage.StorageImages;
+import dev.vitrail.render.timing.PassTimings;
 import dev.vitrail.settings.PackFile;
 import dev.vitrail.settings.PackSession;
 import dev.vitrail.settings.SettingsFile;
