+++ common/src/main/java/dev/vitrail/render/PackImages.java
@@ -1,12 +1,13 @@
 package dev.vitrail.render;
 
+import dev.vitrail.pack.model.PackTexture;
+import dev.vitrail.pack.model.TextureStage;
 import dev.vitrail.pack.source.ShaderPackSource;
 import dev.vitrail.pack.source.ShaderProperties;
 import dev.vitrail.pack.target.SamplerPlan;
-import dev.vitrail.pack.model.PackTexture;
 import dev.vitrail.pack.texture.PackTextures;
-import dev.vitrail.pack.model.TextureStage;
 import dev.vitrail.pack.texture.VolumeAtlas;
+import dev.vitrail.render.pbr.PbrAtlases;
 import dev.vitrail.uniform.NoiseTexture;
 
 import com.mojang.blaze3d.GpuFormat;
