+++ common/src/main/java/dev/vitrail/render/TerrainDraw.java
@@ -6,6 +6,7 @@ import dev.vitrail.pack.source.OpenedPack;
 import dev.vitrail.pack.source.ShadowCasters;
 import dev.vitrail.pack.target.ChainPlan;
 import dev.vitrail.pack.target.TargetPlan;
+import dev.vitrail.render.timing.PassTimings;
 import dev.vitrail.Vitrail;
 
 import com.mojang.blaze3d.pipeline.RenderPipeline;
